fully populate chaco.overlays.api and chaco.api
