+++ chaco/api.py
@@ -56,8 +56,8 @@ Mappers
 - :class:`~.BandedMapper`
 - :class:`~.PolarMapper`
 
-Visual Components
------------------
+Visual Components / Overlays
+----------------------------
 
 - :class:`~.AbstractPlotRenderer`
 - :class:`~.AbstractOverlay`
@@ -71,15 +71,27 @@ Visual Components
 - :class:`~.VPlotContainer`
 - :class:`~.GridPlotContainer`
 - :class:`~.Label`
-- :class:`~.PlotLabel`
-- :class:`~.Legend`
-- :class:`~.ToolTip`
+- :class:`~.ColorBar`
+- :class:`~.AlignedContainerOverlay`
+- :class:`~.ColormappedSelectionOverlay`
+- :class:`~.ContainerOverlay`
+- :class:`~.CoordinateLineOverlay`
+- :class:`~.DataBox`
 - :class:`~.DataLabel`
 - :class:`~.LassoOverlay`
-- :class:`~.ColorBar`
-- :class:`~.TextBoxOverlay`
+- :class:`~.AbstractCompositeIconRenderer`
+- :class:`~.CompositeIconRenderer`
+- :class:`~.PlotLabel`
 - :class:`~.ScatterInspectorOverlay`
-- :class:`~.ColormappedSelectionOverlay`
+- :func:`~.basic_formatter`
+- :func:`~.datetime_formatter`
+- :func:`~.date_formatter`
+- :class:`~.SimpleInspectorOverlay`
+- :func:`~.time_formatter`
+- :class:`~.TextBoxOverlay`
+- :class:`~.TextGridOverlay`
+- :class:`~.ToolTip`
+- :class:`~.ImageInspectorOverlay`
 
 Renderers
 ---------
@@ -294,7 +306,7 @@ from .transform_color_mapper import TransformColorMapper
 from .horizon_plot import BandedMapper
 from .polar_mapper import PolarMapper
 
-# Visual components
+# Visual components / Overlays
 from .abstract_plot_renderer import AbstractPlotRenderer
 from .abstract_overlay import AbstractOverlay
 from .base_plot_container import BasePlotContainer
@@ -319,15 +331,30 @@ except ImportError:
     pass
 
 from .label import Label
-from .plot_label import PlotLabel
-from .legend import Legend
-from .tooltip import ToolTip
-from .data_label import DataLabel
-from .lasso_overlay import LassoOverlay
 from .color_bar import ColorBar
-from .text_box_overlay import TextBoxOverlay
-from .scatter_inspector_overlay import ScatterInspectorOverlay
-from .colormapped_selection_overlay import ColormappedSelectionOverlay
+
+from chaco.overlays.api import (
+    AlignedContainerOverlay,
+    ColormappedSelectionOverlay,
+    ContainerOverlay,
+    CoordinateLineOverlay,
+    DataBox,
+    DataLabel,
+    LassoOverlay,
+    AbstractCompositeIconRenderer,
+    CompositeIconRenderer,
+    PlotLabel,
+    ScatterInspectorOverlay,
+    basic_formatter,
+    datetime_formatter,
+    date_formatter,
+    SimpleInspectorOverlay,
+    time_formatter,
+    TextBoxOverlay,
+    TextGridOverlay,
+    ToolTip,
+    ImageInspectorOverlay
+)
 
 # Renderers
 from .barplot import BarPlot
